[DEPL-121] Refactoring
The key used to store the App and the Main components in Consul does not
end anymore with a "/". Consequently, the consul UI display correctly
the matching value.

A new utility function is also added to record everything into Consul
and launching the application build. This method should be called by
plugins when they start up.